move manager code to autoexit component
